+++ packages/common/src/collections.ts
@@ -126,15 +126,10 @@ const other: string[] = [
 const site: string[] = ["Hub Site Application", "Site Application"];
 
 /**
- * ```js
- * import { getCollection } from "@esri/hub-common";
- * //
- * getCollection('Feature Layer')
- * > 'dataset'
- * ```
  * Get the Hub collection for a given item type
  * @param itemType The ArcGIS [item type](https://developers.arcgis.com/rest/users-groups-and-items/items-and-item-types.htm).
  * @returns the Hub collection of a given item type.
+ * @private
  */
 export const getCollection = (type?: string) => {
   if (!type) {
@@ -147,8 +142,8 @@ export const getCollection = (type?: string) => {
   });
 };
 
-// TODO: remove this at the next breaking change
-// it's only here to support deprecated categories
+// TODO: remove this when we remove the deprecated categories
+// and then move the above arrays and getCollection() logic to get-family
 export const collections: { [key: string]: string[] } = {
   app,
   dataset,