+++ packages/common/src/search/_internal/index.ts
@@ -1,5 +1,4 @@
 export * from "./searchEntities";
-export * from "./state";
 export * from "./portalSearchItems";
 export * from "./hubSearchItems";
 export * from "./portalSearchGroups";