+++ packages/common/src/search/types/index.ts
@@ -1,9 +1,5 @@
-export * from "./ICatalog";
-export * from "./ICollection";
-export * from "./IFacet";
-export * from "./IFacetOption";
 export * from "./IHubAggregation";
-export * from "./IHubApiSearchRequest";
+
 export * from "./IHubSearchOptions";
 export * from "./IHubSearchResult";
 export * from "./IHubSearchResponse";