+++ packages/sites/src/share-items-to-site-groups.ts
@@ -4,14 +4,12 @@ import { _getSecondPassSharingOptions } from "./_get-second-pass-sharing-options
 import { _getSharingEligibleModels } from "./_get-sharing-eligible-models";
 
 /**
- * Share all the other models to the Site's content and collaboration groups, if
- * those groups were created for the site (depends on user's privs)
  * **DEPRECATED: Use shareItemsToSiteGroups() instead**
- * @param {object} siteModel Site Model
- * @param {Array} solutionModels Array of all models created by the Solution
- * @param {IHubRequestOptions} hubRequestOptions
+ * this is only needed in order to support solutions.js < v1.4.1
+ * and we currently use v1.1.5 in opendata-ui
  * @private
  */
+/* istanbul ignore next - deprecated */
 export function _shareItemsToSiteGroups(
   siteModel: IModel,
   solutionModels: IModel[],
@@ -19,10 +17,11 @@ export function _shareItemsToSiteGroups(
 ) {
   /* tslint:disable no-console */
   console.info(
-    `DEPRECATED: _shareItemsToSiteGroups will be removed at v10.0.0. Use shareItemsToSiteGroups instead.`
+    `DEPRECATED: _shareItemsToSiteGroups will be removed at the next breaking version. Use shareItemsToSiteGroups instead.`
   );
   return shareItemsToSiteGroups(siteModel, solutionModels, hubRequestOptions);
 }
+
 /**
  * Share all the other models to the Site's content and collaboration groups, if
  * those groups were created for the site (depends on user's privs)