+++ packages/sites/test/_share-items-to-site-groups.test.ts
@@ -1,9 +1,9 @@
-import { _shareItemsToSiteGroups } from "../src";
+import { shareItemsToSiteGroups } from "../src";
 import * as portalModule from "@esri/arcgis-rest-portal";
 import { IModel, IHubRequestOptions } from "@esri/hub-common";
 import * as sharingEligibleModelsModule from "../src/_get-sharing-eligible-models";
 
-describe("_shareItemsToSiteGroups", () => {
+describe("shareItemsToSiteGroups", () => {
   it("shares items to groups using shareItemsToSiteGroups", async () => {
     const shareSpy = spyOn(portalModule, "shareItemWithGroup").and.returnValue(
       Promise.resolve({})
@@ -34,7 +34,7 @@ describe("_shareItemsToSiteGroups", () => {
       "_getSharingEligibleModels"
     ).and.returnValue(Promise.resolve(toShare.slice(1, 3)));
 
-    await _shareItemsToSiteGroups(siteModel, toShare, hubRequestOptions);
+    await shareItemsToSiteGroups(siteModel, toShare, hubRequestOptions);
 
     expect(_getSharingEligibleModelsSpy.calls.count()).toBe(
       1,