+++ packages/teams/src/utils/get-team-template.ts
@@ -9,12 +9,8 @@ import { getTeamsAvailableInProduct } from "./get-teams-available-in-product";
  * @param {string} product basic | premium | portal
  * @param {string} portalApiVersion portal version
  */
-export function getTeamTemplate(
-  type: HubTeamType,
-  product: HubProduct,
-  portalApiVersion: string
-) {
-  return getTeamsAvailableInProduct(product, portalApiVersion).find(t => {
+export function getTeamTemplate(type: HubTeamType, product: HubProduct) {
+  return getTeamsAvailableInProduct(product).find((t) => {
     return t.config.type === type;
   });
 }