+++ umd-base-profile.js
@@ -1,3 +1,7 @@
+// NOTE: we are no longer running the UMD build
+// we are keeping the profiles, package scripts and dev dependencies
+// in case we discover that we need them later
+// TODO: remove all of the above once we are sure we don't need them
 import resolve from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
 import json from "@rollup/plugin-json";
@@ -75,7 +79,7 @@ arcgisRestJsPackageNames.reduce((globals, p) => {
  * Now we can export the Rollup config!
  */
 export default {
-  // NOTE: you must run the ESM ES5 build before running the UMD build
+  // NOTE: you must run the ESM build before running the UMD build
   input: "./dist/esm/index.js",
   output: {
     file: `./dist/umd/${name.replace("@esri/hub-", "")}.umd.js`,