+++ packages/discussions/CHANGELOG.md
@@ -1,4 +1,10 @@
-## @esri/hub-discussions [11.36.2](https://github.com/Esri/hub.js/compare/@esri/hub-discussions@11.36.1...@esri/hub-discussions@11.36.2) (2022-08-02)
+# @esri/hub-discussions [12.0.0-next.7](https://github.com/Esri/hub.js/compare/@esri/hub-discussions@12.0.0-next.6...@esri/hub-discussions@12.0.0-next.7) (2022-08-03)
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.7
+
+# @esri/hub-discussions [12.0.0-next.6](https://github.com/Esri/hub.js/compare/@esri/hub-discussions@12.0.0-next.5...@esri/hub-discussions@12.0.0-next.6) (2022-08-02)
 
 
 
@@ -6,7 +12,81 @@
 
 ### Dependencies
 
+* **@esri/hub-common:** upgraded to 10.0.0-next.6
+
+# @esri/hub-discussions [12.0.0-next.5](https://github.com/Esri/hub.js/compare/@esri/hub-discussions@12.0.0-next.4...@esri/hub-discussions@12.0.0-next.5) (2022-08-01)
+
+
+### chore
+
+* **hub-common:** no longer publish es5 build ([995ce02](https://github.com/Esri/hub.js/commit/995ce02373e8390250e4490b738babb71cb7c303))
+
+
+### BREAKING CHANGES
+
+* **hub-common:** no longer publish es5 build
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.5
+
+# @esri/hub-discussions [12.0.0-next.4](https://github.com/Esri/hub.js/compare/@esri/hub-discussions@12.0.0-next.3...@esri/hub-discussions@12.0.0-next.4) (2022-08-01)
+
+
+### chore
+
+* **hub-common:** no longer run and publish UMD build ([2f0d7a2](https://github.com/Esri/hub.js/commit/2f0d7a25332e0864e03f814a34847abb8ee1bc4b))
+
+
+### BREAKING CHANGES
+
+* **hub-common:** no longer publish a CDN release
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.4
+
+# @esri/hub-discussions [12.0.0-next.3](https://github.com/Esri/hub.js/compare/@esri/hub-discussions@12.0.0-next.2...@esri/hub-discussions@12.0.0-next.3) (2022-07-28)
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.3
+
+# @esri/hub-discussions [12.0.0-next.2](https://github.com/Esri/hub.js/compare/@esri/hub-discussions@12.0.0-next.1...@esri/hub-discussions@12.0.0-next.2) (2022-07-22)
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.2
+
+# @esri/hub-discussions [12.0.0-next.1](https://github.com/Esri/hub.js/compare/@esri/hub-discussions@11.36.0...@esri/hub-discussions@12.0.0-next.1) (2022-07-21)
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.1
+
+## @esri/hub-discussions [11.36.2](https://github.com/Esri/hub.js/compare/@esri/hub-discussions@11.36.1...@esri/hub-discussions@11.36.2) (2022-08-02)
+
+### Dependencies
+
 * **@esri/hub-common:** upgraded to 9.49.2
+
 
 ## @esri/hub-discussions [11.36.1](https://github.com/Esri/hub.js/compare/@esri/hub-discussions@11.36.0...@esri/hub-discussions@11.36.1) (2022-07-26)