+++ packages/teams/CHANGELOG.md
@@ -1,8 +1,77 @@
+# @esri/hub-teams [10.0.0-next.7](https://github.com/Esri/hub.js/compare/@esri/hub-teams@10.0.0-next.6...@esri/hub-teams@10.0.0-next.7) (2022-08-03)
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.7
+
+# @esri/hub-teams [10.0.0-next.6](https://github.com/Esri/hub.js/compare/@esri/hub-teams@10.0.0-next.5...@esri/hub-teams@10.0.0-next.6) (2022-08-02)
+
+
+### Bug Fixes
+
+* add more orgs with 700 group max ([#838](https://github.com/Esri/hub.js/issues/838)) ([e90a01c](https://github.com/Esri/hub.js/commit/e90a01c583209b32990bfbc79234b39d2147bd7a))
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.6
+
+# @esri/hub-teams [10.0.0-next.5](https://github.com/Esri/hub.js/compare/@esri/hub-teams@10.0.0-next.4...@esri/hub-teams@10.0.0-next.5) (2022-08-01)
+
+
+### chore
+
+* **hub-common:** no longer publish es5 build ([995ce02](https://github.com/Esri/hub.js/commit/995ce02373e8390250e4490b738babb71cb7c303))
+
+
+### BREAKING CHANGES
+
+* **hub-common:** no longer publish es5 build
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.5
+
+# @esri/hub-teams [10.0.0-next.4](https://github.com/Esri/hub.js/compare/@esri/hub-teams@10.0.0-next.3...@esri/hub-teams@10.0.0-next.4) (2022-08-01)
+
+
+### chore
+
+* **hub-common:** no longer run and publish UMD build ([2f0d7a2](https://github.com/Esri/hub.js/commit/2f0d7a25332e0864e03f814a34847abb8ee1bc4b))
+
+
+### BREAKING CHANGES
+
+* **hub-common:** no longer publish a CDN release
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.4
+
+# @esri/hub-teams [10.0.0-next.3](https://github.com/Esri/hub.js/compare/@esri/hub-teams@10.0.0-next.2...@esri/hub-teams@10.0.0-next.3) (2022-07-28)
+
+
+### Code Refactoring
+
+* **hub-teams:** remove portalApiVersion argument from teams functions ([1132c43](https://github.com/Esri/hub.js/commit/1132c436e9f1145f63188b806a9638be0d258c53))
+
+
+### BREAKING CHANGES
+
+* **hub-teams:** remove portalApiVersion argument from teams functions
+
 ## @esri/hub-teams [9.49.3](https://github.com/Esri/hub.js/compare/@esri/hub-teams@9.49.2...@esri/hub-teams@9.49.3) (2022-08-02)
-
-
-
-
 
 ### Dependencies
 
@@ -16,7 +85,27 @@
 
 ### Dependencies
 
-* **@esri/hub-common:** upgraded to 9.49.1
+* **@esri/hub-common:** upgraded to 10.0.0-next.3
+
+# @esri/hub-teams [10.0.0-next.2](https://github.com/Esri/hub.js/compare/@esri/hub-teams@10.0.0-next.1...@esri/hub-teams@10.0.0-next.2) (2022-07-22)
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.2
+
+# @esri/hub-teams [10.0.0-next.1](https://github.com/Esri/hub.js/compare/@esri/hub-teams@9.49.0...@esri/hub-teams@10.0.0-next.1) (2022-07-21)
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 10.0.0-next.1
 
 ## @esri/hub-teams [9.49.1](https://github.com/Esri/hub.js/compare/@esri/hub-teams@9.49.0...@esri/hub-teams@9.49.1) (2022-07-22)
 
@@ -24,6 +113,11 @@
 ### Bug Fixes
 
 * add more orgs with 700 group max ([#838](https://github.com/Esri/hub.js/issues/838)) ([e90a01c](https://github.com/Esri/hub.js/commit/e90a01c583209b32990bfbc79234b39d2147bd7a))
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 9.49.1
 
 # @esri/hub-teams [9.49.0](https://github.com/Esri/hub.js/compare/@esri/hub-teams@9.48.0...@esri/hub-teams@9.49.0) (2022-07-21)